Release/1.0.1

+++ Paltarumi.Acopio.Balanza.Dto/Balanza/Ticket/TicketDto.cs
@@ -8,26 +8,26 @@ namespace Paltarumi.Acopio.Balanza.Dto.Balanza.Ticket
         public DateTimeOffset FechaIngreso { get; set; }
         public DateTimeOffset? FechaSalida { get; set; }
         public int IdEstadoTmh { get; set; }
-        public float PesoBruto100 { get; set; }
-        public float PesoBrutoBase { get; set; }
-        public float PesoBruto { get; set; }
+        public decimal PesoBruto100 { get; set; }
+        public decimal PesoBrutoBase { get; set; }
+        public decimal PesoBruto { get; set; }
         public bool PesoBrutoEdit { get; set; }
         public int? IdUsuarioAprobadorPesoBruto { get; set; }
-        public float Tara { get; set; }
-        public float PesoNeto100 { get; set; }
-        public float PesoNetoBase { get; set; }
-        public float PesoNeto { get; set; }
+        public decimal Tara { get; set; }
+        public decimal PesoNeto100 { get; set; }
+        public decimal PesoNetoBase { get; set; }
+        public decimal PesoNeto { get; set; }
         public int IdEstadoTmhCarreta { get; set; }
-        public float PesoBrutoCarreta100 { get; set; }
-        public float PesoBrutoCarretaBase { get; set; }
-        public float PesoBrutoCarreta { get; set; }
+        public decimal PesoBrutoCarreta100 { get; set; }
+        public decimal PesoBrutoCarretaBase { get; set; }
+        public decimal PesoBrutoCarreta { get; set; }
         public bool PesoBrutoCarretaEdit { get; set; }
         public int? IdUsuarioAprobadorPesoBrutoCarreta { get; set; }
-        public float TaraCarreta { get; set; }
-        public float PesoNetoCarreta100 { get; set; }
-        public float PesoNetoCarretaBase { get; set; }
-        public float PesoNetoCarreta { get; set; }
-        public float PesoNetoTotal { get; set; }
+        public decimal TaraCarreta { get; set; }
+        public decimal PesoNetoCarreta100 { get; set; }
+        public decimal PesoNetoCarretaBase { get; set; }
+        public decimal PesoNetoCarreta { get; set; }
+        public decimal PesoNetoTotal { get; set; }
         public string? Grr { get; set; }
         public string? Grt { get; set; }
         public int? IdTransporte { get; set; }
